Fixed histogram normalization and replaced 1-vs-1 SVM with 1-vs-all

+++ src/Histogram.cpp
@@ -4,7 +4,7 @@ using namespace std;
 Histogram::Histogram() {
 }
 
-Histogram::Histogram(float* data, unsigned int length) {
+Histogram::Histogram(double* data, unsigned int length) {
 	copy(data, data + length, back_inserter(m_data));
 	m_length = length;
 }
@@ -13,6 +13,6 @@ unsigned int Histogram::getLength() const {
 	return m_length;
 }
 
-const float* Histogram::getData() const {
+const double* Histogram::getData() const {
 	return &m_data[0];
 }
